update sleep timer for disconnet
Discord complaints of last.ckpt not appear may be due to auto disconnect only waiting 3 seconds after training this updates it to 40 which is more then enough time to save to the Gdrive (takes about 28 seconds). This will be safe to accept with out further testing due to only timer being changed.